+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-l# Changelog
+# Changelog
 
 <!--next-version-placeholder-->